requirements.txt
python-telegram-bot>=22.0
firebase-admin==6.5.0
python-dotenv==1.0.0
